Fix some of the Tiny£MCE behaviour
