+++ tinymce-per-block/src/blocks/text-block/form.js
@@ -2,27 +2,17 @@
  * External dependencies
  */
 import { createElement } from 'wp-elements';
-import { unescape as unescapeString, reduce, map } from 'lodash';
+import { reduce } from 'lodash';
+import { EditableComponent } from 'wp-blocks';
 
-function getChildElements( child ) {
-	if ( child.value ) {
-		return unescapeString( child.value );
-	}
+import { serialize } from 'serializers/block';
 
-	let children;
-	if ( child.children ) {
-		children = map( child.children, getChildElements );
-	}
-
-	if ( 'HTML_Tag' === child.type ) {
-		return createElement( child.name, child.attrs, children );
-	}
-
-	return children;
-}
-
-export default function TextBlockForm( { node } ) {
-	const style = reduce( node.attrs, ( memo, value, key ) => {
+export default function TextBlockForm( block, { setChildren } ) {
+	const { children } = block;
+	const onChangeContent = ( value ) => {
+		setChildren( value );
+	};
+	const style = reduce( block.attrs, ( memo, value, key ) => {
 		switch ( key ) {
 			case 'align':
 				memo[ 'text-align' ] = value;
@@ -33,11 +23,8 @@ export default function TextBlockForm( { node } ) {
 	}, {} );
 
 	return (
-		<p
-			contentEditable
-			style={ style }
-			className="text-block__form">
-			{ map( node.children, getChildElements ) }
-		</p>
+		<div classnName="text-block__form" style={ style }>
+			<EditableComponent initialContent={ serialize( children ) } onChange={ onChangeContent } />
+		</div>
 	);
 }
